dom/workers/test/test_worker_in_background.html

<html>

<head>
  <title>Workers background flag test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">
var TESTS_RESULTS = new Map();
let worker = new Worker("test_worker_in_background.js");
worker.onmessage = (e) => {
    TESTS_RESULTS.set(e.data.stage, e.data.status);
};
var newtab;

add_task(async function enable_testing_prefs() {
    await SimpleTest.waitForFocus();
    await SpecialPowers.pushPrefEnv({
        set: [
            ["dom.workers.testing.enabled", true],
            ["dom.serviceWorkers.testing.enabled", true],
        ],
    });
    ok(true, "testing settings enabled");
});

add_task(async function test_worker_in_foreground() {
    const stage = "CheckIsForeground";
    worker.postMessage(stage);
    await SimpleTest.promiseWaitForCondition(() => {
        return TESTS_RESULTS.get(stage) != undefined;
    }, stage);
    is(TESTS_RESULTS.get(stage), "PASS", "Worker is in Foreground");
});

add_task(async function test_worker_in_background() {
    const stage = "CheckIsBackground";
    newtab = window.open('about:blank', '_blank');
    document.onvisibilitychange = () => {
        if (document.visibilityState === "hidden") {
            worker.postMessage(stage);
        }
    };
    await SimpleTest.promiseWaitForCondition(() => {
        return TESTS_RESULTS.get(stage) != undefined;
    }, stage);
    is(TESTS_RESULTS.get(stage), "PASS", "Worker is in Background");
});

add_task(async function test_worker_return_to_foreground() {
    const stage = "CheckIsForegroundAgain";
    newtab.close();
    document.onvisibilitychange = () => {
        if (document.visibilityState === "visible") {
            worker.postMessage(stage);
        }
    }
    await SimpleTest.promiseWaitForCondition(() => {
        return TESTS_RESULTS.get(stage) != undefined;
    }, stage);
    is(TESTS_RESULTS.get(stage), "PASS", "Worker is in Foreground again");
});

</script>
</head>



<body>
</body>
</html>
